scsi: Fixup logging under high load
When a lot of messages are to be printed out any SCSI CDBs tend
to come out one byte at a line, which seriously impedes debugging.
This patch uses a local buffer for CDBs to ensure they are
printed with one printk statement which will always be printed
as one line.

Signed-off-by: Hannes Reinecke <hare@suse.de>