scsi: add 'abort_completions' host template flag
Some HBAs will always send a command completion on the original
command after a ABORT TASK TMF has been send.
For those it's pointless to wait for the TMF itself to finish,
but we can as well reset the timer for the failed command and
wait for the command completion to finish up the command.

Signed-off-by: Hannes Reinecke <hare@suse.com>