aha152x: drop bus reset
The bus reset function is a subset of host reset, but will always
return SUCCESS. But as the host reset also invokes a reset for
all ports there is a possibility that the bus reset will not be
able to fully reset the HBA, leading to a blocked system.
So drop bus reset completely.

Signed-off-by: Hannes Reinecke <hare@suse.com>