virtio_scsi: Implement 'native LUN' feature
The 'native LUN' feature allows virtio-scsi to pass in the LUN
numbers from the underlying storage directly, without having
to modify the LUN number itself.
It works by shifting the existing LUN number down by 2 bytes,
and add the virtio-specific 2-byte LUN steering header.
With that virtio doesn't have to mangle the LUN number, allowing
us to pass the 'real' LUN number to the guest.
Of course, we do limit the 'real' LUN number to 6 bytes, but I'm
not aware of any array utilizing more than 4 bytes, so the impact
should be negligible.

Signed-off-by: Hannes Reinecke <hare@suse.com>