aic7xxx: Do not overload 'result' with driver-specific error codes
The aic7xxx driver overloads the 'result' field from struct scsi_cmnd
with driver-specific error codes, which will then be translated back
into the standard error codes upon return.
This is prone to leak the internal error codes and will cause issues
when changing the definitions for the 'result' code.
So use a separate field 'cam_status' in the driver-private command
structure here.

Signed-off-by: Hannes Reinecke <hare@suse.com>