enic: check numa node of page allocated
Pages for all buffers are allocated during initialization. Initialization
process could run on cpu. After interface is up, user can change the cpu
affinity of the interrupt associated with napi. In this case, driver
should discard any pages allocated on non-local memory.

This patch records the numa node and checks if locality changed. If
page locality changes, we discard the page and allocate new page.

If offset of current nc has passt the frag size needed, that means this
buffer is the first buffer which tries to reuse the page. Which means
no other buffer in driver is currently using the frag and we can free the
page.

Signed-off-by: Govindarajulu Varadarajan <gvaradar@cisco.com>