lpfc: Refactor Send LS Response support
Currently, the ability to send an NVME LS response is limited to the nvmet
(controller/target) side of the driver.  In preparation of both the nvme
and nvmet sides supporting Send LS Response, rework the existing send
ls_rsp and ls_rsp completion routines such that there is common code that
can be used by both sides.

Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Paul Ely <paul.ely@broadcom.com>