nvme/fcloop: avoid deadlock during disconnect
When removing a targetport FC-NVMe requires to send a disconnect ELS.
Upon reception of the response the fcloop_targetport_delete() is called,
which calls flush_work() to ensure that all LS responses will be sent.
But as the initial DISCONNECT LS response is sent from the very same
queue we're deadlocking as the workqueue function hasn't completed.
So to avoid this we should return an error straightaway when trying
to send LS responses whenever the remoteport has been disconnected.

Signed-off-by: Hannes Reinecke <hare@suse.de>