nvme: stop mad sqsize increasing and decreasing
The NVMe spec only states:

"One slot in each queue is not available for use due to Head and
Tail entry pointer definition."

It does not say that the actual queue size needs to decremented
by one, it just means that we'll have to leave one slot free.
So kill the mad decreasing and increasing of sqsize, leave it
at the specified size, and only decrement the tagset by one to
ensure we leave one slot free.

Signed-off-by: Hannes Reinecke <hare@suse.de>