aacraid: avoid setting message byte on completion
The aacraid controller is a RAID controller, and the driver will
never see any SCSI messages. Plus it's quite pointless to set
the message byte if the host byte is already set, as the latter
takes precedence during error recovery.
So drop the message byte values for the final result.

Signed-off-by: Hannes Reinecke <hare@suse.de>