scsi: stop setting DRIVER_SENSE
DRIVER_SENSE was meant to signal whether sense data is available.
Which really is quite pointless as one can look at the sense data
to figure that one out.
So drop the usage of DRIVER sense and check the sense buffer
directly when required. And fudge back the DRIVER_SENSE setting
for SG_IO so as not to break userspace.

Signed-off-by: Hannes Reinecke <hare@suse.de>