xen-scsiback: use DID_ERROR instead of DRIVER_ERROR
DRIVER_ERROR was supposed to be set from within the driver, but
xen-scsiback is more like the actual (emulated) hardware, and should
not set DRIVER_ERROR. Plus there's an error where xen-scsiback
wants to return DID_NO_CONNECT, but shifts the value by 24 instead
of the expected 16, causing an invalid (and possibly unhandled)
driver error to be set.
This patch uses DID_ERROR instead of DRIVER_ERROR to
make it more consistent with other drivers and the actual
error handling.

Signed-off-by: Hannes Reinecke <hare@suse.de>