scsi: introduct scsi_result_set_good() helper
There are myriad ways on how one could initialize the 'result'
field of struct scsi_cmnd. And in doing so it's not quite obvious
to the reader whether only the mentioned fields should be set
or indeed the entire result.
So introduce a helper 'scsi_result_set_good()' to make the code
easier to read, and apply a coccinelle script to convert all
drivers to use it.

Signed-off-by: Hannes Reinecke <hare@suse.de>