aacraid: drop AAC_OWNER_LOWLEVEL
AAC_OWNER_LOWLEVEL signals if the associated SCSI command is handed
over from the midlayer to the driver.
But as we have moved owner tracking to struct fib this value does no
longer apply (as we won't have a fib to start with).
So count all SCSI commands with no associated fibs as 'lowlevel'.

Signed-off-by: Hannes Reinecke <hare@suse.de>