af_tlsh: allocate priorities string
The caller of tls_client_hello() can just pass us any string in the
'priorities' argument, and there is no guarantee that the 'priorities'
argument is valid after the call to tls_client_hello().
So we need to allocate a buffer for the priorities string to ensure
it's valid for the lifetime of the socket.

Signed-off-by: Hannes Reinecke <hare@suse.de>